Improve twitter module and add twitterified string
