+++ src/themes/icmaa-imp/components/core/blocks/CategoryExtras/TwitterStatusFeed.vue
@@ -1,37 +1,14 @@
 <template>
   <div>
-    <div v-for="(s, i) in status" :key="i">
-      {{ s.text }}
-    </div>
+    <div v-for="(s, i) in status" :key="i" v-html="twitterify(s.text)" />
   </div>
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
-import { registerModule, isModuleRegistered } from '@vue-storefront/core/lib/modules'
-import { IcmaaTwitterModule } from 'icmaa-twitter'
+import TwitterStatusFeed from 'icmaa-twitter/components/TwitterStatusFeed'
 
 export default {
   name: 'TwitterStatusFeed',
-  props: {
-    screenName: {
-      type: String,
-      required: true
-    }
-  },
-  beforeCreate () {
-    registerModule(IcmaaTwitterModule)
-  },
-  async mounted () {
-    return this.$store.dispatch('icmaaTwitter/loadStatusFeed', this.screenName)
-  },
-  computed: {
-    ...mapGetters({
-      getStatus: 'icmaaTwitter/getStatusByScreenName'
-    }),
-    status () {
-      return this.getStatus(this.screenName)
-    }
-  }
+  mixins: [ TwitterStatusFeed ]
 }
 </script>
